requirements.txt
fastapi
uvicorn[standard]
aiohttp
duckduckgo_search
trafilatura
httpx
python-dotenv
pydantic-settings
playwright